Poll for success of migrations during deployment
The `cf run-task` that we use to run migrations is asynchronous.
This means that currently the pipeline never knows if the migration step
completed successfully, and will continue to ship code to production
even if the migration fails.

This adds a step that polls the CloudFoundry API to check the state of
the migrate task, and returns a non-zero exit code if the state is not
"SUCCEEDED".  This should make the pipeline fail if the migration fails,
so broken migrations will hopefully not make it to production.