Add sequence to Green Deal Charges and Measures
The arrays of Green Deal charges and lists are supposed to correspond - i.e. charges[0] relates to measures[0], charges[1] relates to measures[1], etc.

In the current XML implementation this is guaranteed because XML sequences preserve ordering.  However, JSON does not guarantee consistency of ordering so we need to specify a sequence number to be able to tie measures and charges together.