Update 'status' enum to the values required by Green Deal register
The statuses that the Green Deal register deals with from Landmark's register responses
are "Not for issue" and "Entered".

This is superfluous in our API because an assessment that is "Not for issue" will respond
with 410 anyway, and everything else is "Entered".  But having this field there will
apparently help for Green Deal integration.