Lodge to /assessments NOT /assesments/{assessmentId}
Following consultation with the schemes we have decided on the following way forward:

We will change the lodgement API route (which is used for *all* assessment types) to no longer specify the RRN as a parameter in the URL - what was a POST to `/assessments/{assessmentId}` will instead become a POST to `/assessments` instead.
This route will accept a single <Report> for domestic SAP and RdSAP lodgements, as is currently the case.
For Non-domestic EPC and Air Conditioning it will accept a single XML payload with two <Report> elements - one for the certificate and one for the accompanying report.  These will have different RRNs and be linked as "related reports", as currently happens with the Landmark register.  It will not accept lodgement of one without the other.
For DEC and DEC-RR it will accept a single DEC or DEC-RR, or both together.