Remove "as-is" Open API specification document
up to now we have maintained two versions of the Open API
specification document - one "aspirational" and one "as-is".

This has been useful to be able to signal to integrators
which parts of the API are "ready to test", but has also
caused some confusion.  Additionally, the "as-is" has not
been kept up-to-date as well as it might have been.

At this point in the development the additional Open API
document i more of a hindreance than a help, so we're going
to remove the "as-is" version and refer everyone to the main
api.yml for their integration.