Release any active connections to the connection pool on http response
Active record defaults to holding connections as active outside of the
db connection pool.

Active record, when in a rails app, would usually use the request
lifecycle handler in actionpack https://rubygems.org/gems/actionpack to
ensure request threads give the connections back to the pool on a http
response.

As we are a sinatra app, this capability isn’t available and each thread
has to poll for a connection from the pool.

I think this can be fixed by manually releasing connections when we give
a http response.