Revert commit that changed address identifier pattern
Reverts https://github.com/communitiesuk/epb-register-api/commit/e0caa24f3058b6b10fec8b33017d06b5a7eb9a93

The schema was agreed to be finalised with BRE and schemes, and we
should only make changes now in updated versioned releases.

The pattern here is not ideal (we now know we will always have
12 digits for a UPRN, zero-padding shorter UPRNs).  But this will
be a change for a future version of the schema.