Fix bug where some DEC summaries have empty elements
The problematic DECs have all been lodged with the same piece of
software, which generates XML that has explicit namespaces on every
element, e.g. <CEPC:Report-Type>, <CEPC:RRN> and so on.
This is fine in terms of validating against the schema and getting
lodged, but it seems that when you push this XML through a view
model Nokogiri fails to match anything with e.g.
xpath(%w[Report-Type]) or xpath(%w[RRN]).

The DEC summary relies on the view model to extract values from the
XML, so is currently producing empty elements for DECs where the XML
has explicit CEPC namespace in it.

By preprocessing the DEC XML with remove_namespaces! we can ensure
that summary XML is populated for DECs that were lodged with or
without explicit namespaces.

This commit also introduces the "Additional-Fixtures" folder in
the test samples.  This makes it easier to test a variety of XML
documents for particular cases (or for multiple cases), rather than
only having a single sample for each schema version that must be
used for all tests.