BUGFIX: Address IDs should always be allowed to be updated to the "RRN-" of the certificate itself
In the case where an assessment has been wrongly linked to another RRN-/UPRN- address identifier we need to ba able to set the address identifier to its own RRN if there is no other suitable ID to link it to.

This would previously have resulted in a 400 error, due to mistakenly "mismatched" RRN- identifier.