BUGFIX: <Improvement-Title> is not a thing
Our view models all try to extract from the <Improvement-Title>
element, but this does not exist in any schema, so the /summary
endpoint used by the frontend always had nil for improvementTitle.

This changes the API to instead use the value from <Improvement-Heading>
for the title if it exists, and fall back to <Improvement-Summary> if
it doesn't.

This is the expected behaviour described in the XSD files that define
these elements for SAP and RdSAP.