All domestic certificates on an address are related (SAP & RdSAP)
Current logic is that only certificates of the exact same type are
grouped as related certificates for a property.  However, for
domestic EPCs both SAp and RdSAP certificates should be listed as
"other certificates for this property".

This adds a condition that if two certificates are both either
SAP or RdSAP then they are included in the list of related
certificates. Previously the type had to be strictly identical
for a certificate to be included.