EPBR-1236: Implementing Timecop for all future failing tests
In this commit, we added a Timecop conditional flag into the spec_helper, so tests that use lodge_assessment or any other methods that implement validation rules won't break in 18 months (for domestic EPCs) and 4 years (for non-domestic EPCs). 

A lot of the files in this commit will involve just a flag added to the top of every parent describe block (for purposes of performance -- we don't want to reinstantiate Timecop for every individual test that was failing). 

In the case of get_assessment_rrns_by_scheme_name_and_type_spec.rb, we needed to separately add its own custom Timecop configuration with its own time, because lodgements lodged at midnight 21st June using UTC would have a lodgement_date of 23h00 20th June due to the 1 hour difference when the clocks change to BST. Tests that relied on the lodgement date would then have an incorrect value.

Co-Authored-By: Tony<tony@madetech.com>