Adjust addressId for lodgement with RRN-based addressId
Also updated the set up for AddressSearch::ByBuildingReference spec so the RRN
based address_id doesn't match any existing assessment ids. This is necessary
as our new lodgement method modifies the address_id to match the one for the
existing assessment and consequently links them up.

Now the assessments...0000 and ...0001 are linked together with the address_id
of "RRN-0000-0000-0000-0000-0000".  This address_id is used for the
address_search_by_id request.

Updated the search by address id tests to be more descriptive.

Co-authored-by: Caitlin <caitlin@madetech.com>