Use safe navigation when calculating Green Deal savings
We have a single Green Deal record with a zero "fuel_code".
This currently causes an API error when trying to calculate
the Green Deal savings to display on this certificate, as
there is no corresponding entry in the fuel price table.

By adding safe navigation `&` operators in the calculation
code this should prevent the error and be able to display
the certificate page OK.

NOTE: this is not supposed to be a "feature", but is defensive
programming to help us deal with bad legacy data on this one
certificate.