BUGFIX: Only cancel linked cert if it isn't already cancelled
We've had a report of an uncanellable CEPC.  Turns out this is
because the corresponding CEPC-RR has already been cancelled
individually (before we also cancelled linked certificates).

The existing implementation always attempts to cancel both
linked certificates, but if one is already cancelled then
this results in a "this certificate is already cancelled"
error message, and the still-live certificate doesn't get
cancelled.

The change here means we now only attempt to cancel a
linked certificate if it isn't already cancelled.