Add a task helper to block some tasks in production
There were several different ways of checking for
"production" across our tasks, and several ways
to quit out of them.

This makes a single consistent way for tasks to check
the environment they are running in, that is stricter
than the previous condition by defining specific
STAGEs that are allowed (ie test, development,
integration and staging), rather than allowing
"anything but production".