Make invalid postcode return INVALID_REQUEST error
Search with invalid postcode used to return 500 Server error, as the exception
raised in the helper was not being caught anywhere. The
UseCase::FindAssessmentsByPostcode::PostcodeNotValid is already handled in the
controller.

The logic of checking the regex was moved to the ValidatePostcodeHelper. It
also adds a test for the helper and refactors it to use class methods.