Allow default DECs to be lodged again
The regulations are contradictory on whether asset-rating only DECs are required. Ruling from above is that they should still be able to be lodged.

This commit also makes a couple of other non-domestic lodgement rules non-overridable (floor area can't be negative and dates can't be in the future).

Also small bugfix to add a quotation mark to a lodgement rule error message.