Fall back to created_by for choosing "latest" cert for RHI
Currently if multiple EPCs are lodged on the same day then the RHI "latest" endpoint returns the highest lexicographically sorted RRN, because there is nothing to distinguish on date.

For unmigrated certificates we have a reliable "created_at" timestamp that can be used to tie-break EPCs with identical registration dates.  We need to keep registration date as the primary sort, as this works for all certificates, included migrated ones.

Multiple EPCs lodged on the same day is an edge case, not a usual situation.

For migrated EPCs lodged on the same day we still have no way to reliably distinguish the "latest" lodged on that day.