Strip whitespace from allowed schemas
Using strip(',') to separate lists of schemas means that if someone inadvertently adds a space between the comma-separated schema names then that space is included in the "allowed schema" list and won't match against the schema versions passed in the content-type header of lodgement requests.

We can be resilient to inadvertent spaces in configuration by stripping whitespace from schema names when they are first split out from the environment variable.

This also changes setup slightly so that the environment variables are only read and processed once, the first time the method is called, rather than for every lodgement.