Tweak API spec for Improve Energy Efficiency service
This decouples the domestic EPC search from the service that will use it, as it could be more generally useful in future.

The fetch-data-by-rrn endpoint is unchanged, as the data in here will depend on the DSA with Improve Energy Efficiency service, and so won't be generally applicable to other clients in future.