Fix DEC expiry edge case for increased building size
We had a tricky edge case come in to helpdesk that revealed a bug with our implementation of "superseded_by" in the case where a building increases in size over the threshold of DEC validity swapping from 10 to 1 year:

"A previous 10 year DEC (https://find-energy-certificate.service.gov.uk/energy-certificate/9900-3940-0136-7711-8020) was lodged to the same UPRN in 2016, when the building was under 1,000 sq.m. Despite the 10 year DEC being lodged in 2016, the system won't recognise the new DEC (https://find-energy-certificate.service.gov.uk/energy-certificate/0508-3312-9712-7102-4892) as the most recent one - this is because the new DEC is an increased sq.m (now over 1000) changing the validity period to annual."

The fix here is to use the registration-date rather than expiry-date to order related assessments.