Allow 18-month age limit for domestic EPCs to be overridden
Elmhurst have an EPC that has failed audit but is over 18 months
old so a replacement EPC can't be lodged.
By making the 18-month rule overridable they will be able to re-
lodge the EPC.

We want to keep the "dates in future" as *not* overridable, so
this commit splits out the DATES_IN_RANGE rule into two parts,
the new DATES_CANT_BE_IN_FUTURE rule mirrors the non-domestic
version of the rule, and DATES_IN_RANGE now only triggers for
things over 18 months old.