Process building_complexity as a number, not a string
The "building_complexity" key for CEPC's is actually
populated with the value from <Building-Level> in the
view models.  This used to be returned as a string, but
in the new view models version it is an integer, so we
can now check this as a number rather than a string.