Fix API to expect integers not strings from view models
building_level is now guaranteed to be an integer, and this
is reflected in the SBEM lodgement rule here.

energy efficiency ratings and environmental impact are also
now integer values from view models.

This commit does not fix the open data tests, which will be
made together in a subsequent commit.