Update Open Data test expectations for integers not strings
The update of our view models Gem means that a number of
fields are now treated as integers rather than strings.

This commit assumes that Open Data will also prefer these
fields in number format, rather than strings. If this is
not the case then we can revert this commit and update
the use case to cast these fields back to string format
before export.