Energy band calculator returns nil band for nil rating
Previously this was called with a "rating.to_i" from the
presenters, which Ruby "helpfully" returns zero for if
the rating is nil, thus giving an "a" band rating for
commercial, or "g" for domestic.

If there's no rating value provided then the band should
also be nil.

This fixes an issue we have seen for DECs where fewer than
three year's data is included.  Before view models 2.0.0
the rating had to_i called with it first, so the nil value
for subsequent year's ratings was converted to a zero(!)
before the band was calculated.  This now properly returns
nil for a nil input.