EPBR-6496 move test functions into shared_context blocks
defining function in an rspec file means the function lives in the globla namespace. This can cause name collision when code is being executed and lead to 'leaky' test. All these functions definitions have been added to a shared_context meaning the scoped only the tests that use said context