EPBR-6704 updates the import postcode rake to raises errors
Previously the check_task_requirements method aborts when arguments are missing. This functionality was not tested. Updates the check_task_requirements() to raise errors as opposed to aborting. This allows the code to be tested. Additionally adds a rule to raise an error if the file name passed the rake is incorrect