Updating rubocop version
This rubocop update removes rexml as a dependency from rubocop which means it doesn't get built into our app. This commit adds it as an explicit dependency as we use it for parsing xml at lodgement. See incident report 10.

This commit also contains some corrections raised by the updated rubocop.