Remove unused XML read to Hash from lodgement code
We were planning to switch XMLMini to use Nokogiri backend instead
of rexml.  On investigating some errors that were thrown by the
Nokogiri backend (it didn't like xmlns not being an absolute URI)
we discovered that the result of the call to Hash.from_xml(data)
is never actually used anywhere in the code.

So we can remove this line altogether, along with the dependency
on rexml from the Gemfile, because the code is no longer using
XMLMini at all.