Example of is_scottish flag being used to direct lodgements to different schemas.
When the flag is false, all tests pass as normal.

When the flag is true, you can see in this case where I haven't added the database migration yet, that the insert fails because the schema doesn't yet exist. Next example will show result with the new schema in place.