Another example of is_scottish flag being used to direct field updates to different schemas - difference is use of ActiveRecord class in previous one v use of bare sql in this one.
When the flag is false, all tests pass as normal.

When the flag is true, you can see in this case where I haven't added the database migration yet, that the update fails because the schema doesn't yet exist. Next example will show result with the new schema in place.