EPBR-8964 Spike showing how we can run migrations on the scottish schema.
We have a second directory for scottish relevant migrations. Idea is people will copy relevant migrations over into this directory. The `migrate_scotland` task then runs the migrations in that directory against the scottish schema. I've also shown how this task can be hooked on to the normal `db:migrate` task so that both sets of migrations run with the usual command.

To test this you will need to have added the scottish schema to your db - see the jira ticket for this.