Improve some tests
This updates instances where we had tests like `expect(thing1.to_a - thing2.to_a).to eq []` -
sometimes the first item was an empty array and so subtracting from it gave a false pass for
the test. I've prefered to compare hashes were possible, and if not possible compare the
array in both directions.

This has led to some expectations needing updating (I suspect the code has changed to remove
fields etc over time, but as the test looked like it was passing the expectation wasn't
updated).

There is a leaky lodgement in open_data_export_spec.rb - it was affecting the "returns the
csv values with no line breaks" test which is why I've pulled the specific row out
(`cleaned_up_data`) so the tests pass, but it would be worth looking into why
parsed_exported_data sometimes contains an extra lodgement from within this test file.