EPBR-9018 Update the migrate lodgement process to add Scottish RdSAP assessments to the Scotland schema
Updated the foreign keys in the Scotland schema tables to point to tables in the correct schema (not always the Scotland schema)

Gone through the lodge_ assessment use case. I have added in an is_scottish flag which can then be passed into the dependant methods and gateways.

I have only altered things which are necessary for migration not lodgement. I have added #TODO flags where things will need to be changed later.

In the gateways I have added the is_scottish flag which then allows the gateway to swap between schemas.

I have added tests for Scottish assessments in the gateway unit tests and end to end acceptance tests for lodging a Scottish assessment.